nhrpd: consider IP version on tunnel addr. change
Not taking the IP version into account was for instance leading to
cancel and re-schedule the transmission of a registration request with
an IPv4 destination protocol address when an IPv6 address was assigned
to the tunnel. This change has a potential impact on the use of IPSec
encryption on a GRE IPv6 tunnel that must be investigated.