snapcraft: rename snap and add extra version info
The snap name is changed to 'iotr-frr' to allow publishing to the Snap
Store. Additionnally, an extra version is appended to the FRR release
number (7.5.1) to track changes in Itron’s fork of the project. For
the extra version information to be appended, execute the GNU
configure script with the package extra version option

    git clean -xfd
    ./bootstrap.sh
    ./configure --with-pkg-extra-version=`cat snapcraft/extra_version_info.txt`
    make dist
    cd snapcraft && snapcraft --destructive-mode
    sudo snap install --devmode iotr-frr_7.5.1-0.1.0_amd64.snap