nhrpd: use a pseudo-device as tunnel source
Add specific handling for the case where a pseudo-device is configured as a
tunnel source interface. A TUN device for instance would disappear then reappear
when the corresponding user process is restarted. On disappearance, the source
device of the GRE tunnel would be set to a bogus value reflecting the link index
of the disappeared device (e.g. if42) and remain as is unless reconfigured from
nhrpd.