nhrpd: customize NBMA iface IPv6 address selection
In order to handle Itron Gen5 RF mesh access point change scenario. In
that scenario up to 3 global IPv6 addresses can be assigned to the RF
mesh interface of a DMVPN spoke. Upon the addition of an IPv6 default
route, the gateway address is compared against the prefix of all the
addresses assigned to the RF mesh interface and the matching address,
if any, is selected.